refactor(domain): move use cases to specific packages
