+++ core/domain/src/test/java/com/casecode/pos/core/domain/usecase/business/AddBranchBusinessUseCaseTest.kt
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.casecode.pos.core.domain.usecase
+package com.casecode.pos.core.domain.usecase.business
 
+import com.casecode.pos.core.domain.usecase.AddBranchBusinessUseCase
 import com.casecode.pos.core.domain.utils.AddBranchBusinessResult
 import com.casecode.pos.core.model.data.users.Branch
 import com.casecode.pos.core.testing.repository.TestBusinessRepository
