+++ core/domain/src/test/java/com/casecode/pos/core/domain/usecase/business/GetBusinessUseCaseTest.kt
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.casecode.pos.core.domain.usecase
+package com.casecode.pos.core.domain.usecase.business
 
+import com.casecode.pos.core.domain.usecase.GetBusinessUseCase
 import com.casecode.pos.core.domain.utils.BusinessResult
 import com.casecode.pos.core.model.data.users.Business
 import com.casecode.pos.core.testing.repository.TestBusinessRepository
