+++ core/domain/src/test/java/com/casecode/pos/core/domain/usecase/business/SetBusinessUseCaseTest.kt
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.casecode.pos.core.domain.usecase
+package com.casecode.pos.core.domain.usecase.business
 
 import com.casecode.pos.core.domain.R
 import com.casecode.pos.core.domain.repository.AddBusiness
+import com.casecode.pos.core.domain.usecase.SetBusinessUseCase
 import com.casecode.pos.core.domain.utils.Resource
 import com.casecode.pos.core.model.data.users.Branch
 import com.casecode.pos.core.model.data.users.Business
