+++ core/domain/src/test/java/com/casecode/pos/core/domain/usecase/business/SetSubscriptionBusinessUseCaseTest.kt
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.casecode.pos.core.domain.usecase
+package com.casecode.pos.core.domain.usecase.business
 
 import com.casecode.pos.core.domain.R
+import com.casecode.pos.core.domain.usecase.SetSubscriptionBusinessUseCase
 import com.casecode.pos.core.domain.utils.Resource
 import com.casecode.pos.core.model.data.users.SubscriptionBusiness
 import com.casecode.pos.core.testing.repository.TestSubscriptionsBusinessRepository
