+++ core/domain/src/test/java/com/casecode/pos/core/domain/usecase/item/UpdateItemUseCaseTest.kt
@@ -13,48 +13,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.casecode.pos.core.domain.usecase
+package com.casecode.pos.core.domain.usecase.item
 
+import com.casecode.pos.core.domain.usecase.UpdateItemUseCase
 import com.casecode.pos.core.domain.utils.Resource
 import com.casecode.pos.core.model.data.users.Item
 import com.casecode.pos.core.testing.repository.TestItemRepository
 import com.casecode.pos.core.testing.util.MainDispatcherRule
 import kotlinx.coroutines.test.runTest
 import org.junit.Rule
-import kotlin.test.Test
+import org.junit.Test
 import kotlin.test.assertEquals
 import com.casecode.pos.core.data.R.string as stringData
 
-class AddItemUseCaseTest {
+class UpdateItemUseCaseTest {
     @get:Rule
     val mainDispatcherRule = MainDispatcherRule()
 
     // Subjects under test
     private val testItemRepository = TestItemRepository()
-    private val addItemUseCase = AddItemUseCase(testItemRepository)
+    private val updateItemUseCase = UpdateItemUseCase(testItemRepository)
 
     @Test
-    fun addItem_whenItemAdded_returnsSuccess() = runTest {
-        // Given
-        val newItem =
-            Item(
-                name = "New Item",
-                unitPrice = 10.0,
-                quantity = 22,
-                sku = "1212312",
-                imageUrl = "newItemImage",
-            )
-        // When
-        val result = addItemUseCase(newItem)
-        // Then
-        assertEquals(
-            result,
-            (Resource.success(stringData.core_data_item_added_successfully)),
-        )
-    }
-
-    @Test
-    fun addItem_whenHasError_returnsError() = runTest {
+    fun whenHasError_returnsError() = runTest {
         // Given
         testItemRepository.setReturnError(true)
         val newItem =
@@ -66,8 +47,31 @@ class AddItemUseCaseTest {
                 imageUrl = "newItemImage",
             )
         // When
-        val result = addItemUseCase(newItem)
+        val result = updateItemUseCase(newItem)
         // Then
-        assertEquals(result, Resource.error(stringData.core_data_add_item_failure_generic))
+        assertEquals(
+            (Resource.error(stringData.core_data_update_item_failure_generic)),
+            result
+        )
+    }
+
+    @Test
+    fun whenItemUpdated_returnsSuccess() = runTest {
+        // Given
+        val newItem =
+            Item(
+                name = "New Item",
+                unitPrice = 10.0,
+                quantity = 22,
+                sku = "121231   2",
+                imageUrl = "newItemImage",
+            )
+        // When
+        val result = updateItemUseCase(newItem)
+        // Then
+        assertEquals(
+            result,
+            (Resource.success(stringData.core_data_item_updated_successfully)),
+        )
     }
 }
